move main func into packge so it can be tested
